[HOT-FIX] Removed warning about dtype getting set to object by default. Added note to analyzeMC and rerunMC scripts informing user to first run the vanilla MC script first.
Former-commit-id: c74dc0548a5b70e9d11e6592f3a20c8324425043
Former-commit-id: 419a7a083ac05447c685014ad6bd88e6c46bc4cd

+++ src/utilities/datashader_utilities.py
@@ -33,7 +33,7 @@ def curve_per_df_component(df):
         varIdx_df = df.loc[idx[:], idx[:, i]]
 
         # Inject NaNs at the end of the run so the curves don't wrap from t_f to t_0
-        varIdx_df = varIdx_df.append(pd.Series(name=np.nan))
+        varIdx_df = varIdx_df.append(pd.Series(name=np.nan, dtype='float'))
 
         # Flatten values by column order
         time = np.tile(varIdx_df.index, len(varIdx_df.columns.codes[0]))  # Repeat time by number of runs
